# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Scalp AI Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PriceResponse {
  /** Current XAUUSD price */
  price: number;
  /** Price change from previous close */
  change: number;
  /** Price change percentage */
  changePercent: number;
  /** 24-hour high */
  high24h: number;
  /** 24-hour low */
  low24h: number;
  /** ISO timestamp of price data */
  timestamp: string;
}

export interface SessionInfo {
  london: boolean;
  newYork: boolean;
  asian: boolean;
  active: string;
}

/**
 * Scalp trading signal
 */
export type SignalResponseSignal =
  (typeof SignalResponseSignal)[keyof typeof SignalResponseSignal];

export const SignalResponseSignal = {
  LONG: "LONG",
  SHORT: "SHORT",
  HOLD: "HOLD",
} as const;

/**
 * Overall market trend
 */
export type SignalResponseTrend =
  (typeof SignalResponseTrend)[keyof typeof SignalResponseTrend];

export const SignalResponseTrend = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

/**
 * Signal strength classification
 */
export type SignalResponseSignalStrength =
  | (typeof SignalResponseSignalStrength)[keyof typeof SignalResponseSignalStrength]
  | null;

export const SignalResponseSignalStrength = {
  STRONG: "STRONG",
  NORMAL: "NORMAL",
} as const;

/**
 * 5m EMA trend
 */
export type IndicatorsTrend5m =
  (typeof IndicatorsTrend5m)[keyof typeof IndicatorsTrend5m];

export const IndicatorsTrend5m = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

/**
 * 1m EMA trend (entry confirmation)
 */
export type IndicatorsTrend1m =
  (typeof IndicatorsTrend1m)[keyof typeof IndicatorsTrend1m];

export const IndicatorsTrend1m = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

export interface Indicators {
  /** RSI 14 on 5m */
  rsi: number;
  /** EMA 9 on 5m */
  ema9: number;
  /** EMA 21 on 5m */
  ema21: number;
  macdLine: number;
  macdSignal: number;
  macdHistogram: number;
  /** ATR 14 on 5m */
  atr: number;
  /** 5m EMA trend */
  trend5m: IndicatorsTrend5m;
  /** 1m EMA trend (entry confirmation) */
  trend1m: IndicatorsTrend1m;
}

export interface SignalResponse {
  /** Scalp trading signal */
  signal: SignalResponseSignal;
  /** Confidence score 0-100 */
  confidence: number;
  /** Recommended entry price */
  entryPrice: number;
  /** Stop loss price */
  stopLoss: number;
  /** Take profit price */
  takeProfit: number;
  /** Overall market trend */
  trend: SignalResponseTrend;
  /** Human-readable reason for the signal */
  reason: string;
  /** ISO timestamp of signal generation */
  timestamp: string;
  /** Expected trade duration e.g. 5-30 minutes */
  tradeDuration: string;
  /** Seconds remaining in cooldown period */
  cooldownRemaining: number;
  /** Whether Smart Mode is currently active */
  smartMode: boolean;
  /** Signal strength classification */
  signalStrength: SignalResponseSignalStrength;
  /** EMA crossover score 0-35 */
  emaScore: number;
  /** RSI score 0-25 */
  rsiScore: number;
  /** MACD score 0-25 */
  macdScore: number;
  /** 1m momentum confirmation score 0-15 */
  momentumScore: number;
  /** Fair Value Gap entry zone score 0-25 */
  fvgScore: number;
  /** Liquidity sweep score 0-30 */
  sweepScore: number;
  session: SessionInfo;
  indicators: Indicators;
}

export type HistoryEntrySignal =
  (typeof HistoryEntrySignal)[keyof typeof HistoryEntrySignal];

export const HistoryEntrySignal = {
  LONG: "LONG",
  SHORT: "SHORT",
  HOLD: "HOLD",
} as const;

export type HistoryEntryTrend =
  (typeof HistoryEntryTrend)[keyof typeof HistoryEntryTrend];

export const HistoryEntryTrend = {
  BULLISH: "BULLISH",
  BEARISH: "BEARISH",
  NEUTRAL: "NEUTRAL",
} as const;

/**
 * Trade outcome status
 */
export type HistoryEntryTradeStatus =
  (typeof HistoryEntryTradeStatus)[keyof typeof HistoryEntryTradeStatus];

export const HistoryEntryTradeStatus = {
  RUNNING: "RUNNING",
  TARGET_HIT: "TARGET_HIT",
  STOP_HIT: "STOP_HIT",
  HOLD: "HOLD",
} as const;

export interface HistoryEntry {
  id: string;
  signal: HistoryEntrySignal;
  confidence: number;
  entryPrice: number;
  stopLoss: number;
  takeProfit: number;
  trend: HistoryEntryTrend;
  reason: string;
  timestamp: string;
  tradeDuration: string;
  /** Trade outcome status */
  tradeStatus: HistoryEntryTradeStatus;
  /** Price at which trade was closed (TP or SL hit) */
  closedPrice?: number;
  /** ISO timestamp when trade was closed */
  closedAt?: string;
  /** Profit or loss in USD points */
  pnlPoints?: number;
}

export interface HistoryResponse {
  signals: HistoryEntry[];
}

export interface PositionSizeRequest {
  /** Account balance in USD */
  balance: number;
  /** Risk percentage (e.g., 1 for 1%) */
  riskPercent: number;
  /** Stop loss distance in USD (price difference) */
  stopLossDistance: number;
}

export interface PositionSizeResponse {
  /** Calculated lot size */
  lotSize: number;
  /** Risk amount in USD */
  riskAmount: number;
  /** Total position value in USD */
  positionValue: number;
  /** Value per pip */
  pipValue: number;
}

export interface SmartModeRequest {
  /** Enable or disable Smart Mode */
  enabled: boolean;
}

export type TradeRecordSignal =
  (typeof TradeRecordSignal)[keyof typeof TradeRecordSignal];

export const TradeRecordSignal = {
  LONG: "LONG",
  SHORT: "SHORT",
} as const;

export type TradeRecordResult =
  (typeof TradeRecordResult)[keyof typeof TradeRecordResult];

export const TradeRecordResult = {
  WIN: "WIN",
  LOSS: "LOSS",
} as const;

export interface TradeRecord {
  id: number;
  signal: TradeRecordSignal;
  result: TradeRecordResult;
  entryPrice: number;
  closedPrice: number;
  pnlPoints: number;
  timestamp: string;
}

export type AnalyticsResponseRecentTrend =
  (typeof AnalyticsResponseRecentTrend)[keyof typeof AnalyticsResponseRecentTrend];

export const AnalyticsResponseRecentTrend = {
  HOT: "HOT",
  COLD: "COLD",
  STABLE: "STABLE",
  LEARNING: "LEARNING",
} as const;

export interface AnalyticsResponse {
  /** Total completed trades */
  totalCompleted: number;
  wins: number;
  losses: number;
  /** Win rate as a percentage (0-100) */
  winRate: number;
  lossRate: number;
  /** Average profit on winning trades (USD points) */
  avgProfit: number;
  /** Average loss on losing trades (USD points) */
  avgLoss: number;
  /** Expected value per trade */
  expectancy: number;
  last10: TradeRecord[];
  smartMode: boolean;
  smartModeStatus: string;
  learningStatus: string;
  sufficientData: boolean;
  /** Current win/loss streak (positive = wins, negative = losses) */
  streak: number;
  recentTrend: AnalyticsResponseRecentTrend;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ClearHistory200 = {
  deleted: number;
};

export type DeleteSignal200 = {
  success: boolean;
};