# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gold Scalp AI Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history/{id}:
    delete:
      operationId: deleteSignal
      tags: [trading]
      summary: Delete a signal record
      description: Permanently deletes a signal from history by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "404":
          description: Signal not found
...
  /analytics/smart-mode:
    post:
      operationId: setSmartMode
      tags: [trading]
      summary: Toggle Smart Mode
      description: Enable or disable strict signal filtering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmartModeRequest"
      responses:
        "200":
          description: Smart mode updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
...
components:
  schemas:
...
    SmartModeRequest:
      type: object
      properties:
        enabled:
          type: boolean
          description: Enable or disable Smart Mode
      required:
        - enabled
    TradeRecord:
      type: object
      properties:
        id:
          type: number
        signal:
          type: string
          enum: [LONG, SHORT]
        result:
          type: string
          enum: [WIN, LOSS]
        entryPrice:
          type: number
        closedPrice:
          type: number
        pnlPoints:
          type: number
        timestamp:
          type: string
      required:
        - id
        - signal
        - result
        - entryPrice
        - closedPrice
        - pnlPoints
        - timestamp
    AnalyticsResponse:
      type: object
      properties:
        totalCompleted:
          type: number
          description: Total completed trades
        wins:
          type: number
        losses:
          type: number
        winRate:
          type: number
          description: Win rate as a percentage (0-100)
        lossRate:
          type: number
        avgProfit:
          type: number
          description: Average profit on winning trades (USD points)
        avgLoss:
          type: number
          description: Average loss on losing trades (USD points)
        expectancy:
          type: number
          description: Expected value per trade
        last10:
          type: array
          items:
            $ref: "#/components/schemas/TradeRecord"
        smartMode:
          type: boolean
        smartModeStatus:
          type: string
        learningStatus:
          type: string
        sufficientData:
          type: boolean
        streak:
          type: number
          description: Current win/loss streak (positive = wins, negative = losses)
        recentTrend:
          type: string
          enum: [HOT, COLD, STABLE, LEARNING]
      required:
        - totalCompleted
        - wins
        - losses
        - winRate
        - lossRate
        - avgProfit
        - avgLoss
        - expectancy
        - last10
        - smartMode
        - smartModeStatus
        - learningStatus
        - sufficientData
        - streak
        - recentTrend
...