# lib/api-zod/src/generated/types/tradeRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Scalp AI Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRecordResult } from "./tradeRecordResult";
import type { TradeRecordSignal } from "./tradeRecordSignal";

export interface TradeRecord {
  id: number;
  signal: TradeRecordSignal;
  result: TradeRecordResult;
  entryPrice: number;
  closedPrice: number;
  pnlPoints: number;
  timestamp: string;
}